requirements.txt
black==25.1.0
django==5.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
drf-spectacular==0.28.0
psycopg2-binary==2.9.10
uvicorn[standard]==0.34.2
